requirements.txt
Pillow~=10.0.1
PyQt5~=5.15.4
requests~=2.31.0
image~=1.5.33